Corregido: "Issue 28:	Actualizar listas invertidas en los index records" Al actualizar un record del indice invertido ya no pisa la lista invertida anterior. Se modificó el InvertedIndexRecord para que al obtener la lista invertida devuelva una copia. Se modificó DocumentRepository.java para que al guardar un TextDocument le setee el id. Las busquedas funcionean correctamente.

+++ trunk/src/speakit/ftrs/InvertedListGenerator.java
@@ -47,7 +47,7 @@ public class InvertedListGenerator {
 			return new InvertedList();
 		}
 		int frecuency = 0;
-		int currentDoc = appearanceList.get(0).getDocument();
+		long currentDoc = appearanceList.get(0).getDocument();
 		Occurrence app = null;
 		for (Iterator<Occurrence> appIterator = appearanceList.iterator(); appIterator.hasNext();) {
 			app = appIterator.next();
